cleanup
warnings
prettier
divided nft confirm to two components
tags styles to separate file
constants from copmonents to folder
