+++ public/locales/en.json
@@ -267,23 +267,20 @@
   "auctionEnds": " Auctions Ends In",
   "zeroBalanceMessage": " You don't have balance",
   "zeroBalance": " No balance",
-   "Days":"Days",
-  "Hours":"Hours",
-  "Minutes":"Minutes",
-  "delistAsset":"Delisting Asset",
-  "delistCost":"Delisting the asset from the market place right now Will cost you stakes if it's didn't Exceed the minimaum duration of delisting",
-  "delistNow":"Delisting the asset now means it will be removed from the marketplace are you sure you want to delist it?",
-  "minDelistLeft":" minimum delisting duration Left",
-  "delistConfirm":" I’m sure about delisting my asset which will charge me stake fees",
-  "delistNowButton":" Delist Now",
-  "cancelDelisting":"Cancel Delisting",
-  "DelistSuccess":"Your asset has been Delisted and moved to “Inventory, Off marketplace”",
-  "checkStake":" Check my Stake Balance",
-  "checkInventory":"Check my Inventory",
-  "BackToMarketplace":"Go Back to Marketplace",
-  "DelistWarning":"You Asset cannot be delisted Until the Auction Time Ends",
-  "DelistBidWarning":"Your Asset cannot be delisted because it’s in auction and already got bids"
-
-
-
+  "Days": "Days",
+  "Hours": "Hours",
+  "Minutes": "Minutes",
+  "delistAsset": "Delisting Asset",
+  "delistCost": "Delisting the asset from the market place right now Will cost you stakes if it's didn't Exceed the minimaum duration of delisting",
+  "delistNow": "Delisting the asset now means it will be removed from the marketplace are you sure you want to delist it?",
+  "minDelistLeft": " minimum delisting duration Left",
+  "delistConfirm": " I’m sure about delisting my asset which will charge me stake fees",
+  "delistNowButton": " Delist Now",
+  "cancelDelisting": "Cancel Delisting",
+  "DelistSuccess": "Your asset has been Delisted and moved to “Inventory, Off marketplace”",
+  "checkStake": " Check my Stake Balance",
+  "checkInventory": "Check my Inventory",
+  "BackToMarketplace": "Go Back to Marketplace",
+  "DelistWarning": "You Asset cannot be delisted Until the Auction Time Ends",
+  "DelistBidWarning": "Your Asset cannot be delisted because it’s in auction and already got bids"
 }
