+++ src/components/Button/index.tsx
@@ -375,7 +375,6 @@ export const ButtonSearch = styled(ButtonPrimary)`
   width: 8.5vw;
   border-radius: 0px 4px 4px 0px;
   font-size: 1.125rem;
-
 `
 
 interface ButtonMintProps {
