+++ src/components/Header/MarketplaceHeader.tsx
@@ -4,16 +4,9 @@ import Logo from '../../assets/icons/logo.svg'
 import { ButtonSearch } from 'components/Button'
 import { LinkCreateNFT } from 'components/Link'
 import { InputSearch } from 'components/Input/styles'
-import Books from '../../assets/icons/bookstab.svg'
-import Videos from '../../assets/icons/videostab.svg'
-import Art from '../../assets/icons/arttab.svg'
-import Games from '../../assets/icons/gamestab.svg'
-import All from '../../assets/icons/alltab.svg'
-import Music from '../../assets/icons/musictab.svg'
-import Images from '../../assets/icons/imagestab.svg'
 import { useGetNFTs } from 'state/marketplace/hooks'
 import { useHistory } from 'react-router'
-import { CATEGORIES, DEFAULT_SORT, Dictionary, HEADER_DROPDOWN } from '../../constants'
+import { ALL_CATEGORIES, DEFAULT_SORT, HEADER_DROPDOWN, TabIcons } from '../../constants'
 import { useTranslation } from 'react-i18next'
 import { useWalletAddress } from 'state/user/hooks'
 import { useLocationSearch } from 'hooks'
@@ -23,18 +16,6 @@ import { useDeposit } from 'hooks/startfiStakes'
 import { useWeb3React } from '@web3-react/core'
 import { useApproveToken } from 'hooks/startfiToken'
 import { address as STARTFI_STAKES_ADDRESS } from '../../constants/abis/StartfiStakes.json'
-
-const Categories = ['all', ...CATEGORIES]
-
-const TabIcons: Dictionary = {
-  books: Books,
-  videos: Videos,
-  art: Art,
-  games: Games,
-  all: All,
-  music: Music,
-  images: Images
-}
 
 const MarketplaceHeader: React.FC = () => {
   const address = useWalletAddress()
@@ -102,7 +83,7 @@ const MarketplaceHeader: React.FC = () => {
       {!address && <ConnectWallet>{t('marketplaceConnectWallet')}</ConnectWallet>}
 
       <TabsCategory>
-        {Categories.map(c => (
+        {ALL_CATEGORIES.map(c => (
           <Tab key={c} selected={category === c} onClick={() => history.push(`/marketplace/nfts?category=${c}`)}>
             <Img src={TabIcons[c]} alt={c} />
             {t(c)}
