+++ src/components/MintNFT/Step2.tsx
@@ -17,7 +17,7 @@ const Step2: React.FC<StepProps> = ({ state, handleChange, missing }: StepProps)
       />
 
       <Tags name="tags" max={10} onChange={handleChange} value={state.tags} />
-      
+
       <Input
         name="description"
         label="NFTdescription"
