+++ src/components/NFTConfirm/index.tsx
@@ -1,21 +1,13 @@
-import ButtonWishlist from 'components/Button/ButtonWishlist'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { useHistory } from 'react-router-dom'
 import { usePopup } from 'state/application/hooks'
-import { useAuctionNFT, useBidOrBuy, useBidOrBuyValue, useBuyNFT, usePlaceBid } from 'state/marketplace/hooks'
-import { useUserBalance } from 'state/user/hooks'
-import { shortenAddress } from 'utils'
+import { useAuctionNFT } from 'state/marketplace/hooks'
 import uriToHttp from 'utils/uriToHttp'
+import Payment from './Payment'
 import {
   Bold,
-  Border,
-  ButtonBlack,
-  ButtonTransparent,
-  MarginLeft,
-  Right,
   SemiBold,
-  SpaceBetween,
   Baseline,
   Text,
   Container,
@@ -36,19 +28,7 @@ const NFTConfirm: React.FunctionComponent = () => {
 
   const history = useHistory()
 
-  const balance = useUserBalance()
-
   const auctionNFT = useAuctionNFT()
-
-  const placebid = usePlaceBid()
-
-  const buynft = useBuyNFT()
-
-  const bidOrBuy = useBidOrBuy()
-
-  const v: any = useBidOrBuyValue()
-
-  const value = parseFloat(v)
 
   const popup = usePopup()
 
@@ -58,24 +38,7 @@ const NFTConfirm: React.FunctionComponent = () => {
     return null
   }
 
-  if (value === 0) {
-    popup({ success: false, message: 'noValue' })
-    history.goBack()
-    return null
-  }
-
   const { ownername, issuername, ownerdetails, nft } = auctionNFT
-
-  const service = (): number => value * 1
-
-  const total = (value: any, service: any): number => {
-    return parseFloat(value) + parseFloat(service)
-  }
-
-  const confirm = () => {
-    if (bidOrBuy) placebid()
-    else buynft()
-  }
 
   return (
     <Container>
@@ -116,39 +79,7 @@ const NFTConfirm: React.FunctionComponent = () => {
         </MarginBottom>
         <TextBlack>{ownerdetails}</TextBlack>
       </Left>
-      <MarginLeft>
-        <Right minHeight="74vh">
-          <Bold>{t(bidOrBuy ? 'confirmBidding' : 'confirmPayment')}</Bold>
-          <TextBlack>{t('bidDesc', { ownername, owner: shortenAddress(nft.owner, 6) })}</TextBlack>
-          <SpaceBetween>
-            <Bold>{t(bidOrBuy ? 'biddingOffer' : 'paymentAmount')}</Bold>
-            <Bold>{value} STFI</Bold>
-          </SpaceBetween>
-          <Border />
-          <SpaceBetween>
-            <SemiBold>{t(bidOrBuy ? 'biddingBalance' : 'paymentBalance')}</SemiBold>
-            <Bold>{value} STFI</Bold>
-          </SpaceBetween>
-          <SpaceBetween>
-            <SemiBold>{t('yourBalance')}</SemiBold>
-            <Bold>{balance} STFI</Bold>
-          </SpaceBetween>
-          <SpaceBetween>
-            <SemiBold>{t('serviceFees')}</SemiBold>
-            <Bold>{service()} STFI</Bold>
-          </SpaceBetween>
-          <Border />
-          <SpaceBetween>
-            <Bold>{t(bidOrBuy ? 'totalBidAmount' : 'totalPaymentAmount')}</Bold>
-            <Bold>{total(value, service())} STFI</Bold>
-          </SpaceBetween>
-          <ButtonBlack onClick={() => confirm()}>{t(bidOrBuy ? 'confirmBidding' : 'confirmPayment')}</ButtonBlack>
-          <ButtonWishlist nftId={nft.id} type="NFTConfirm" />
-          <ButtonTransparent onClick={() => history.goBack()}>
-            {t(bidOrBuy ? 'cancelBidding' : 'cancelPayment')}
-          </ButtonTransparent>
-        </Right>
-      </MarginLeft>
+      <Payment />
     </Container>
   )
 }
