+++ src/hooks/useAllowed.tsx
@@ -17,7 +17,6 @@ export const useAllowed = (): boolean => {
           ? getApproverAddress(tokenId).then(approver => setAllowed(approver === (STARTFI_Marketplace_ADDRESS as any)))
           : null
       )
-      return () => {}
   }, [address, nft, getApproverAddress])
   return allowed
 }
