+++ src/pages/Marketplace.tsx
@@ -17,7 +17,7 @@ const Marketplace: React.FC = () => {
   return (
     <MarketplaceWrapper>
       <MarketplaceHeader />
-      <Route path="/marketplace/nfts"  component={NFTs} />
+      <Route path="/marketplace/nfts" component={NFTs} />
       <Route path="/marketplace/nft/:nft/:auction" component={Nftproduct} />
       <Route path="/marketplace/buyorbid" component={NFTConfirm} />
       <Route path="/marketplace/wishList" component={WishList} />
