+++ src/services/Marketplace.tsx
@@ -40,7 +40,7 @@ export const getMarketplace = async (query?: NFTQUERY) => {
   const auctionSort = sort ? sort : DEFAULT_SORT
   const nfts = await getNFTs(nftsQuery)
   const auctions = await getAuctionsPaginated({ status: 'open' }, sortHelper(auctionSort), lastAuction)
-  var onMarket: AuctionNFT[] = []
+  let onMarket: AuctionNFT[] = []
   auctions.forEach((rawAuction: any) => {
     const auction = rawAuction.data()
     const nft = nfts.filter((nft: NFT) => nft.id === auction.nft)[0]
@@ -58,7 +58,7 @@ export const getMarketplace = async (query?: NFTQUERY) => {
   })
   onMarket = onMarket.slice(0, NFTS_PER_PAGE)
   const id = onMarket[onMarket.length - 1]?.auction.id
-  var newLastAuction = null
+  let newLastAuction = null
   auctions.forEach((a: any) => {
     if (a.data().id === id) newLastAuction = a
   })
@@ -67,50 +67,6 @@ export const getMarketplace = async (query?: NFTQUERY) => {
   return { onMarket, loadtime, newLastAuction, auctions, ...query }
 }
 
-// export const getMarketplace = async (query?: NFTQUERY) => {
-//   const t0 = performance.now()
-//   if (!query) query = {}
-//   const { search, category, sort, lastAuction } = query
-//   var chainId
-//   if (!query.chainId) chainId = DEFAULT_CHAIN_ID
-//   else chainId = query.chainId
-//   const nftsQuery: NFTQUERY = { chainId }
-//   if (search) nftsQuery.name = search
-//   if (category && category !== 'all') nftsQuery.category = category
-//   const auctionSort = sort ? sort : DEFAULT_SORT
-//   const nfts = await getNFTs(nftsQuery)
-
-//   const auctions = await getAuctionsPaginated({ status: 'open', chainId }, sortHelper(auctionSort), lastAuction)
-
-//   var onMarket: AuctionNFT[] = []
-//   auctions.forEach((rawAuction: any) => {
-//     const auction = rawAuction
-//     const nft = nfts.filter((nft: NFT) => nft.id === auction.nft)[0]
-
-//     if (nft) {
-//       if (nft.issueDate) delete nft.issueDate
-//       if (auction.purchaseTime) delete auction.purchaseTime
-
-//       onMarket.push({
-//         nft,
-//         auction,
-//         ownername: '',
-//         issuername: '',
-//         ownerdetails: ''
-//       })
-//     }
-//   })
-//   onMarket = onMarket.slice(0, NFTS_PER_PAGE)
-//   const id = onMarket[onMarket.length - 1]?.auction.id
-//   var newLastAuction = null
-//   auctions.forEach((a: any) => {
-//     if (a.id === id) newLastAuction = a
-//   })
-//   const t1 = performance.now()
-//   const loadtime = Math.round(t1 - t0)
-//   return { onMarket, loadtime, newLastAuction, auctions, ...query }
-// }
-
 interface GetAuctionNFT {
   nftId: string
   auctionId: string
@@ -118,8 +74,8 @@ interface GetAuctionNFT {
 }
 
 export const getAuctionNFT = async ({ nftId, auctionId, AuctionNFT }: GetAuctionNFT) => {
-  var nft: NFT
-  var auction: Auction
+  let nft: NFT
+  let auction: Auction
 
   if (AuctionNFT) {
     nft = AuctionNFT.nft
@@ -147,12 +103,11 @@ export const getAuctionNFT = async ({ nftId, auctionId, AuctionNFT }: GetAuction
 interface BuyNFT {
   nftId: number
   auctionId: string
-  owner: string
   buyer: string
   soldPrice: number
 }
 
-export const buyNFT = async ({ nftId, auctionId, owner, buyer, soldPrice }: BuyNFT) => {
+export const buyNFT = async ({ nftId, auctionId, buyer, soldPrice }: BuyNFT) => {
   const nftChange = {
     id: nftId,
     owner: buyer
